# swagger.yaml
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://ingress",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Storage service",
      "description": "Работа с файлами"
    },
    {
      "name": "Statistic service",
      "description": "Работа со статистикой"
    }
  ],
  "paths": {
    "/api/v1/storage": {
      "post": {
        "tags": [
          "Storage service"
        ],
        "description": "Получение ссылки на загруженный файл",
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/admin-console/statistic/post": {
      "get": {
        "tags": [
          "Statistic service"
        ],
        "description": "Получение статистики постов",
        "operationId": "getPostStatistic",
        "parameters": [
          {
            "name": "postStatisticRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostStatisticRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/admin-console/statistic/like": {
      "get": {
        "tags": [
          "Statistic service"
        ],
        "description": "Получение статистики лайков",
        "operationId": "getLikeStatistic",
        "parameters": [
          {
            "name": "PostStatisticRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostStatisticRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/admin-console/statistic/comment": {
      "get": {
        "tags": [
          "Statistic service"
        ],
        "description": "Получение статистики комментариев",
        "operationId": "getCommentStatistic",
        "parameters": [
          {
            "name": "PostStatisticRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostStatisticRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/admin-console/statistic/account": {
      "get": {
        "tags": [
          "Statistic service"
        ],
        "description": "Получение статистики аккаунтов",
        "operationId": "getAccountStatistic",
        "parameters": [
          {
            "name": "accountStatisticRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccountStatisticRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStatisticResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "StorageDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Ссылка на файл в облаке"
          }
        },
        "description": "Dto загруженного файла"
      },
      "PostStatisticRequestDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "firstMonth": {
            "type": "string",
            "format": "date-time"
          },
          "lastMonth": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StatisticPerDateDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "StatisticResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "countPerHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatisticPerDateDto"
            }
          },
          "countPerMonth": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatisticPerDateDto"
            }
          }
        }
      },
      "PostStatisticRequestDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "firstMonth": {
            "type": "string",
            "format": "date-time"
          },
          "lastMonth": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PostStatisticRequestDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "firstMonth": {
            "type": "string",
            "format": "date-time"
          },
          "lastMonth": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AccountStatisticRequestDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "firstMonth": {
            "type": "string",
            "format": "date-time"
          },
          "lastMonth": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AccountCountPerAge": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AccountStatisticResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "countPerAge": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountCountPerAge"
            }
          },
          "countPerMonth": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatisticPerDateDto"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}